Engine: reverted some recent changes to height of lines measurement
This *partially* reverts d9923528b5a0ab7cd4cdb319d35ade868a8ed9fe , which did 2 changes,
* one - adding size of outline when calculating text height (this is kept)
* and another - using full font's graphical extent when calculating height for text windows dialog options (that is reverted).

The latter is reverted, because it causes unnecessary gaps on textwindow objects in previously made games, with the majority of TTF fonts (it seems).
Also I just realized that it made it impossible to calculate future text overlay height using GetTextHeight script command, because this function would have a different result. So it's logically flawed.

Using real graphical height of a font when calculating sizes of objects on screen should be only done with explicit instruction; we currently have a global option that lets do this in new games, and so engine should rely on a result of that option rather than hardcoding "use max font's extent in this calculation".
(Although it's quite possible that we need more adjusting options for rare weird fonts, but that's a separate issue.)