Engine: fixed preparing 8-bit room bgs in hi-color games
Each room background might have its own palette, therefore when converting 8-bit bgs in16-bit or 32-bit color games we must use respective palette for each of the backgrounds.

The 8-bit room bg behavior in 16/32-bit games was broken by commit 4480d98cb5467dc9e4d60d3dc83f86e2be6a42a2. Prior to that the 8-bit images were not converted to game color depth, which made them react to palette changes at runtime, which is not expected in principle when we are running a hicolor game. After that commit they are required to be converted on load. It's essential that a proper palette is selected when we load them up.