Engine: fix < 3.6.0 games may have unintended overhotspot label text
This counters a change done by 23493a681d4f66ae0b1088cddb82241b0bc80dbb as a part of 3.6.0 development.
Older games had a slightly different order of update, noteably overhotspot labels were drawn *before* updating a mouse position, meaning their text would refer to a previous cursor pos.
With the above change some old games sometimes have unintended overhotspot text appearing for a split second where it was not supposed to, *possibly* because it was somehow hidden in the original game's script, but now they could not hide it in time. This is a case with e.g. a old version of "Gemini Rue".