Engine: don't use alpha channel in colored gui controls in 2.*  games
Investigating a strange issue with a 2.7 game "Diamonds in the Rough" (see #2819),
I found by experimentation that 2.72 and earlier engines did not fixup drawing color values with opaque alpha when painting primitives on 32-bit surfaces in 32-bit games.
This includes anything that is not a sprite: raw drawing, but also default gui controls (with no image assigned).
As a result, whenever a GUI background was considered to contain alpha channel (i.e. have a sprite with alpha as a background), then anything painted as simple colored primitive will become fully transparent.

This commit implements a backwards compatible mode for emulating this.