Engine: correct special case for TTF ascender fixup
This corrects/complements commit 14679a430bf29acb6e511bb3799a5f99d9929463
An exception to the ascender fixup was introduced to fix pre-3.4.1 games with the text anti-aliasing option. But it appears that this exception works only in few versions between 3.2.0 and 3.4.1.
Anything before 3.2.0 should not have it, or fonts may be loaded incorrectly.