Engine: fixed auto-outline for fonts with glyphs with negative offsets
Some rare fonts may have glyphs that have negative offsets, that is - their leftmost pixels lie to the left from the "pen" position (requested top-left text render coords).
We may learn this by reading font's BBox.
This commit fixes this problem for the auto-outlining. But the general issue still remains: if such text is positioned too close to the leftmost border of a parent surface (gui control, overlay, etc), then there's a chance that parts of the letter will be cut off.
This issue cannot be resolved without displacing text further from its starting draw position. But that may go against text rendering logic, and produce effects unexpected to the user. So I'll leave that for the future consideration.