Engine: nested script callbacks are run immediately on a temp thread
Change the way how new callbacks triggered by a running script are run. Historically they got scheduled to run on the main "script thread" after the current script finishes, which could be much time later if there are blocking actions or Wait calls in the script.
The new implementation runs such callbacks immediately always. If the main thread is busy, then a dynamic "script thread" gets allocated for each nested callback, with no hard limit on the number of nested callbacks. Such dynamic threads exist only for the callback's duration.

This naturally affects the order of callbacks, so users might have to fix their script (depending on situation). But the advantage of having a deterministic order and time of a callback run makes this worth it IMO.

Additionally, this mechanism opens a potential for supporting function pointers / delegates in script.