Editor: display colors as a single property, that RGB or pal index
This is a partial backport from ags4 branch.
Since early days of 3.0 Editor the colors in game settings were represented by a pair of properties, one of which let to set color as a "number" (palette index, or a 16-bit encoded RGB), or as a (R,G,B) sequence.

This change merges these pairs of properties into one property per setting, which may be initialized either as (R,G,B) in 16/32-bit game, or as a palette index in a 8-bit game.

There's no change to the game data, project source nor compiled data, this is only a change to the user interface.